requirements.txt
langchain==0.0.284
langchain-openai==0.1.0
langchain-community==0.0.10
langchain-anthropic==0.1.0
langchain-google-genai==0.1.0
python-dotenv==1.0.0
streamlit==1.22.0
unstructured==0.9.2
tiktoken==0.4.0
faiss-cpu==1.7.4
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
openai==0.28.0
anthropic==0.18.1
google-generativeai==0.8.3
sentence-transformers==2.5.1
torch==2.2.0
transformers==4.38.2